Add special-case support for Reddit and HN.
The Reddit RSS feed is structured such that the <link> element takes you
to the Reddit discussion thread, while the <description> element
contains an HTML document that contains both a link to Reddit, and a
link to the actual article content itself.  As a result, this plugin is
unable to inline the linked article in the feed (since the <link>
element goes to Reddit, not the article).

Hacker News, by contrast, populates the <link> and <comment> tags
correctly, while also supplying a link to the Hacker News comments in
the <description> element, so the plugin works correctly.  However,
unlike the web interface, the tt-rss mobile app does not provide a way
to navigate to the comment URL (yet).  Because this plugin replaces the
article content, it renders the HN comment link inaccessible on a mobile
device.

To work around these two issues, this changeset does three things for
Reddit or HN RSS feeds:

First, it attempts to pull the link out of the article content.  If it's
there, the url element of the article is replaced with the found link.
The content is then extracted from the replacement URL, which means the
article content is properly inlined into the feed.

Second, it appends a link to the comment thread to the article content
itself, so that even on a mobile device, there's a way to get there.

Third, for transparency, the change prepends the source website URL to
the article description, so that you can see where the content came from
at a glance.